fixed data loading
