+++ packages/kokoas-client/src/components/ui/dialogs/ContentWarning.tsx
@@ -1,21 +1,23 @@
 import { Stack, Typography } from '@mui/material';
 import { ReactNode } from 'react';
 import WarningIcon from '@mui/icons-material/Warning';
+import { StackProps } from '@mui/system';
 
 export const ContentWarning = ({
   content,
-}: {
+  ...stackProps
+}: StackProps & {
   content: ReactNode
 }) => {
   return (
-    <Stack>
+    <Stack {...stackProps} spacing={2}>
       <Typography textAlign={'center'}>
         <WarningIcon
           fontSize={'large'}
           color="warning"
         />
       </Typography>
-      <Typography>
+      <Typography textAlign={'center'}>
         {content}
       </Typography>
     </Stack>
