+++ packages/kokoas-client/src/pages/projRegister/hooks/useResolveParams.ts
@@ -1,10 +1,11 @@
 
 
-import { parseISO, format } from 'date-fns';
+
 import { useProjById, useCustGroupById } from 'kokoas-client/src/hooksQuery';
 import { useEffect, useState } from 'react';
-import { BuildingType, RecordCancelStatus, RecordStatus, TAgents } from 'types';
+
 import { getParam } from '../../../helpers/url';
+import { convertCustGroupToForm, convertProjToForm } from '../api/convertToForm';
 import { TypeOfForm, initialValues } from '../form';
 
 /**
@@ -19,66 +20,33 @@ export const useResolveParams = () => {
   const { data: projRec } = useProjById(projIdFromURL || '');
   const { data: custGroupRec } = useCustGroupById(projRec?.custGroupId.value || custGroupIdFromURL || '');
 
-  useEffect(() => {
-    if (projRec) {
-      const {
-        projTypeId, projName,
-        custGroupId, $id,
-        isAgentConfirmed, postal, address1, address2,
-        buildingType, isChkAddressKari, agents, addressKari,
-        status,
-        cancelStatus,
-        projTypeName,
-        storeId,
-        作成日時: createTime,
-      } = projRec;
-
-      const cocoConst = agents.value.filter(item => {
-        return (item.value.agentType.value as TAgents) === 'cocoConst';
-      }).map(item => item.value.agentId.value);
-
-      setInitForm(prev => ({
-        ...prev,
-        addressKari: addressKari.value,
-        address1: address1.value,
-        address2: address2.value,
-        buildingType: buildingType.value as BuildingType,
-        cancelStatus: cancelStatus.value.split(',') as RecordCancelStatus[],
-        cocoConst1: cocoConst?.[0],
-        cocoConst2: cocoConst?.[1],
-        createdDate: format(parseISO(createTime.value), 'yyyy/MM/dd'),
-        custGroupId: custGroupId.value,
-        isAgentConfirmed: Boolean(+isAgentConfirmed.value),
-        isChkAddressKari: Boolean(+isChkAddressKari.value),
-        projId: $id.value,
-        projTypeId: projTypeId.value,
-        projTypeName: projTypeName.value,
-        projName: projName.value,
-        postal: postal.value,
-        storeId: storeId.value,
-        status: (status?.value as RecordStatus) || '追客中',
-      }));
-    }
-  }, [projRec]);
 
   useEffect(() => {
-    if (custGroupRec) {
-      const {
-        storeId,
-        territory,
-        $id,
-        members,
-      } = custGroupRec;
-      setInitForm((prev) => ({
-        ...prev,
-        custGroupId: $id.value,
-        storeId: storeId.value,
-        territory: territory.value,
-        custName: members.value[0]?.value.customerName.value || '',
-      }));
 
+    if (projIdFromURL) {
+      if (projRec && custGroupRec) {
+        setInitForm({
+          ...initialValues,
+          ...convertProjToForm(projRec),
+          ...convertCustGroupToForm(custGroupRec),
+        });
+      }
     }
-  }, [custGroupRec]);
+
+
+    if (custGroupIdFromURL && !projIdFromURL) {
+      if (custGroupRec) {
+        setInitForm({
+          ...initialValues,
+          ...convertCustGroupToForm(custGroupRec),
+        });
+      }
+    }
+
+    if (!custGroupIdFromURL && !projIdFromURL) {
+      setInitForm(initialValues);
+    }
+  }, [projRec, custGroupRec, projIdFromURL, custGroupIdFromURL ]);
 
   return initForm;
 };
